[feat/SOP-13]: Adding status

+++ src/main/java/com/benevides/sop_api/domain/commitment/CreateCommitmentDTO.java
@@ -3,6 +3,7 @@ package com.benevides.sop_api.domain.commitment;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public record CreateCommitmentDTO(
@@ -10,7 +11,7 @@ public record CreateCommitmentDTO(
         Date date,
 
         @Positive(message = "O valor é obrigatório e deve ser maior que zero")
-        float value,
+        BigDecimal value,
 
         String note,
