+++ src/main/java/com/benevides/sop_api/domain/commitment/GetCommitmentsWithPaymentCountDTO.java
@@ -1,11 +1,12 @@
 package com.benevides.sop_api.domain.commitment;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public record GetCommitmentsWithPaymentCountDTO(
         long id,
         Date date,
-        float value,
+        BigDecimal value,
         String note,
         String commitment_number,
         long payment_count
