+++ src/main/java/com/benevides/sop_api/domain/payment/CreatePaymentDTO.java
@@ -3,6 +3,7 @@ package com.benevides.sop_api.domain.payment;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public record CreatePaymentDTO(
@@ -10,7 +11,7 @@ public record CreatePaymentDTO(
         Date date,
 
         @Positive(message = "O valor é obrigatório e deve ser maior que zero")
-        float value,
+        BigDecimal value,
 
         String note,
